Fix compilation on Android itself
This is clearly not the most common way to build for Android,
but I'm experimenting with building on an Android tablet
(see https://github.com/dfaure/rust-android-hello-world for howto)
and this crate failed to build for lack of an Android implementation
of find_android_sdk().

No default installation directory, so setting ANDROID_HOME is
required.